+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Issue #1497: Fix styling of `featureviaurl` features that was previously missing, resulting in invisible features.
 - Issue #1505: Fix FileImport tool breaking after using ShareView URL with the tool being open.
 - Issue #1509: Fix zIndex order of layers configured by URL that could've been wrong in edge cases.
+- Issue #1512: Fix `compareMaps` for mobile devices, where the slider was previously missing.
 - Issue #1519: Fix initialization of print module.
 - Searchbar and LayerSelection: Behavior restored to the way it was before layerSelection filtering was introduced.